Break apart the send confirmation styles
