+++ src/components/scenes/SendConfirmationScene.js
@@ -4,7 +4,7 @@ import { bns } from 'biggystring'
 import { Scene } from 'edge-components'
 import type { EdgeCurrencyInfo, EdgeCurrencyWallet, EdgeDenomination, EdgeMetadata, EdgeSpendInfo, EdgeTransaction } from 'edge-core-js'
 import React, { Component } from 'react'
-import { TouchableOpacity, View } from 'react-native'
+import { StyleSheet, TouchableOpacity, View } from 'react-native'
 import slowlog from 'react-native-slowlog'
 import { sprintf } from 'sprintf-js'
 
@@ -24,9 +24,9 @@ import { Slider } from '../../modules/UI/components/Slider/Slider.ui.js'
 import { type AuthType, getSpendInfoWithoutState } from '../../modules/UI/scenes/SendConfirmation/selectors'
 import { convertCurrencyFromExchangeRates } from '../../modules/UI/selectors.js'
 import { type GuiMakeSpendInfo, type SendConfirmationState } from '../../reducers/scenes/SendConfirmationReducer.js'
-import { rawStyles, styles } from '../../styles/scenes/SendConfirmationStyle.js'
 import { THEME } from '../../theme/variables/airbitz.js'
 import type { GuiCurrencyInfo, GuiDenomination, GuiWallet } from '../../types/types.js'
+import { scale } from '../../util/scaling.js'
 import { convertNativeToDisplay, convertNativeToExchange, decimalOrZero, getDenomFromIsoCode } from '../../util/utils.js'
 import { AddressTextWithBlockExplorerModal } from '../common/AddressTextWithBlockExplorerModal'
 import { SceneWrapper } from '../common/SceneWrapper.js'
@@ -305,7 +305,7 @@ export class SendConfirmation extends Component<Props, State> {
                   {isTaggableCurrency && (
                     <Scene.Row style={{ paddingVertical: 10 }}>
                       <TouchableOpacity
-                        activeOpacity={rawStyles.activeOpacity}
+                        activeOpacity={THEME.OPACITY.ACTIVE}
                         style={styles.addUniqueIDButton}
                         onPress={this.props.uniqueIdentifierButtonPressed}
                       >
@@ -518,6 +518,75 @@ export class SendConfirmation extends Component<Props, State> {
     this.setState({ fioSender })
   }
 }
+
+const rawStyles = {
+  mainScrollView: {
+    flex: 1,
+    width: '100%',
+    alignItems: 'center'
+  },
+  exchangeRateContainer: {
+    alignItems: 'center',
+    marginVertical: scale(12)
+  },
+
+  main: {
+    alignItems: 'center',
+    width: '100%'
+  },
+  feeAreaText: {
+    fontSize: scale(16),
+    color: THEME.COLORS.WHITE,
+    backgroundColor: THEME.COLORS.TRANSPARENT
+  },
+
+  sliderStyle: {
+    width: scale(270)
+  },
+  error: {
+    marginHorizontal: scale(10),
+    backgroundColor: THEME.COLORS.TRANSPARENT
+  },
+  errorText: {
+    textAlign: 'center',
+    color: THEME.COLORS.ACCENT_RED
+  },
+  balanceText: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(16)
+  },
+  balanceContainer: {
+    alignItems: 'center',
+    marginTop: scale(10)
+  },
+  rowText: {
+    backgroundColor: THEME.COLORS.TRANSPARENT,
+    color: THEME.COLORS.WHITE
+  },
+  pinInputContainer: {
+    width: scale(60),
+    height: scale(50)
+  },
+  pinInputSpacer: {
+    width: scale(10)
+  },
+  addUniqueIDButton: {
+    backgroundColor: THEME.COLORS.TRANSPARENT,
+    padding: scale(14),
+    alignItems: 'center',
+    justifyContent: 'center'
+  },
+  addUniqueIDButtonText: {
+    color: THEME.COLORS.WHITE
+  },
+  footer: {
+    marginTop: scale(12)
+  },
+  footerWithPaymentId: {
+    marginTop: scale(0)
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
 
 export const uniqueIdentifierText = (currencyCode: string, uniqueIdentifier?: string): string => {
   if (!getSpecialCurrencyInfo(currencyCode).uniqueIdentifier) {
